move dialogs to shared folder and fix quicklink dialog
