+++ src/features/table/fileCard/FileTags.tsx
@@ -1,7 +1,7 @@
 import NiceModal from "@ebay/nice-modal-react"
 import { Badge } from "@src/components/ui/badge"
-import ViewTaggedFilesDialog from "@src/features/dialogs/tags/ViewTaggedFilesDialog"
 import { TAGS_QUERY_KEY, getFileTags } from "@src/features/files/tags/storage"
+import ViewTaggedFilesDialog from "@src/features/shared/dialogs/tags/ViewTaggedFilesDialog"
 import { useQuery } from "@tanstack/react-query"
 
 interface FileTagsProps {
